Cooja: Move seeding to platform_init_stage_one
